add RelationOutput symbol and ExpressionAnalyzer
This commit adds a RelationOutput, ExpressionAnalyzer and a
RelationOutputResolver.

The ExpressionAnalyzer can be used to convert Expression
(from SQL AST) into a Symbol. Using composition over
inheritance makes it possible to easier handle nested
relations and re-use the ExpressionAnalyzer for different
sources.

The DataStatementAnalyzer is now marked as deprecated and
shouldn't be used anymore.